Refactor library view and loader into focused modules
